+++ src/primative_examples/primativeexamples.rs
@@ -1,0 +1,9 @@
+
+
+pub fn hello(name: &str) {
+    println!("hello {}", name);
+}
+
+pub fn add(a: i8, b: i8) {
+    println!("{}", a + b);
+}